requirements.txt
keras-cv
keras-nlp
keras
transformers
torch
torchvision
torchaudio
bitsandbytes
datasets
evaluate
peft
scikit-learn
wandb
sagemaker
trl
python-Levenshtein
accelerate
rtree
geopandas
shapely